# openapi/OptimaixMaster/main.yml
...
openapi: 3.0.0
info:
  title: Optimaix Master Node API
  description: TBD
  version: 0.1.0
#  servers: # these are for test servers TBD 
#    - url: http://api.example.com/v1
#      description: Optional server description, e.g. Main (production) server
...
paths:
  /oaasNodes:
    get:
      summary: Returns a list of the registered (federated) OaaS nodes
      description: TBD
      tags:
        - OaaS Nodes
      responses:
        '200':    # status code
          description: A JSON array of user full information (all but password info)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './schemas/OaasNodeOut.yml'
    
    post:
      summary: Registers (federates) a new OaaS node in th system
      description: TBD
      requestBody: 
        content: 
          application/json: 
            schema:
              $ref: './schemas/OaasNodeIn.yml'
      tags:
        - OaaS Nodes
      responses:
        '200':    # status code
          description: All ok
          content:
            application/json:
              schema:
                $ref: './schemas/OaasNodeOut_single.yml'
  /oaasNodes/{oaasNodeId}:
    get:
      summary: Return info of the registered (federated) OaaS Node
      description: TBD
      parameters:
        - $ref: './parameters/oaasNodeId.yml'
      tags:
        - OaaS Nodes
      responses:
        '200':    # status code
          description: TBD
          content:
            application/json:
              schema:
                $ref: './schemas/OaasNodeOut.yml'

    put:
      summary: Adds a new algorithm repository to the OaaS node
      description: TBD
      requestBody: 
        content: 
          application/json: 
            schema:
              $ref: './schemas/OaasNodeUpdate.yml'
      parameters:
        - $ref: './parameters/oaasNodeId.yml'
      tags:
        - OaaS Nodes
      responses:
        '200':    # status code
          description: All ok
          content:
            application/json:
              schema: 
                  $ref: './schemas/OaasNodeOut.yml'
    delete: 
      summary: Deletes an existing algorithm repository
      description: TBD
      parameters:
        - $ref: './parameters/oaasNodeId.yml'
      tags:
        - OaaS Nodes
      responses:
        '200':    # status code
          description: All ok

# expose the full API of the OaaS node, to be callable by the user, via the master
# see the form of having a full list without focusing on a single oaasNodeId
...